long

+++ backend/src/main/java/com/example/backend/entities/Order.java
@@ -91,7 +91,7 @@ public class Order implements Serializable {
         this.status = status;
     }
 
-    public long getPrice() {
+    public double getPrice() {
         return price;
     }
